refactor: rename express loader to app loader and delete useless comments in app.js
