+++ loaders/app.js
